fix Raja function

+++ src/LinAlg/hiopVectorRajaPar.cpp
@@ -511,10 +511,10 @@ void hiopVectorRajaPar::copy_from_two_vec_w_pattern(const hiopVector& c,
   const hiopInt* id2 = ix2.local_data_const();
   
   RAJA::forall< hiop_raja_exec >(
-    RAJA::RangeSegment(0, n1_local),
+    RAJA::RangeSegment(0, (int)n1_local),
     RAJA_LAMBDA(RAJA::Index_type i)
     {
-      dd[id1[i]] = vd1[i];
+      dd[(int)id1[i]] = vd1[i];
     }
   );
 
@@ -522,7 +522,7 @@ void hiopVectorRajaPar::copy_from_two_vec_w_pattern(const hiopVector& c,
     RAJA::RangeSegment(0, n2_local),
     RAJA_LAMBDA(RAJA::Index_type i)
     {
-      dd[id2[i]] = vd2[i];
+      dd[(int)id2[i]] = vd2[i];
     }
   );
 }
@@ -556,18 +556,18 @@ void hiopVectorRajaPar::copy_to_two_vec_w_pattern(hiopVector& c,
   const hiopInt* id2 = ix2.local_data_const();
   
   RAJA::forall< hiop_raja_exec >(
-    RAJA::RangeSegment(0, n1_local),
+    RAJA::RangeSegment(0, (int)n1_local),
     RAJA_LAMBDA(RAJA::Index_type i)
     {
-      vd1[i] = dd[id1[i]];
+      vd1[i] = dd[(int)id1[i]];
     }
   );
 
   RAJA::forall< hiop_raja_exec >(
-    RAJA::RangeSegment(0, n2_local),
+    RAJA::RangeSegment(0, (int)n2_local),
     RAJA_LAMBDA(RAJA::Index_type i)
     {
-      vd2[i] = dd[id2[i]];
+      vd2[i] = dd[(int)id2[i]];
     }
   );                                           
 }
